Add metadata fields on home and blocks pages
