--- packages/frontend/src/routes/dataContracts/data.contracts.route.js
@@ -1,13 +1,0 @@
-import React, {useState, useEffect} from 'react';
-import {Form, useLoaderData} from "react-router-dom";
-import * as Api from "../../util/Api";
-
-function DataContractsRoute() {
-    return (
-        <div className="App">
-            <div>Еasdasdк</div>
-        </div>
-    );
-}
-
-export default DataContractsRoute;
