+++ packages/frontend/src/routes/enums/state.transition.type.js
@@ -1,0 +1,10 @@
+export const StateTransitionEnum = {
+    DATA_CONTRACT_CREATE: 0,
+    DOCUMENTS_BATCH: 1,
+    IDENTITY_CREATE: 2,
+    IDENTITY_TOP_UP: 3,
+    DATA_CONTRACT_UPDATE: 4,
+    IDENTITY_UPDATE: 5,
+    IDENTITY_CREDIT_WITHDRAWAL:6,
+    IDENTITY_CREDIT_TRANSFER: 7
+};
