+++ packages/frontend/src/util/index.js
@@ -1,0 +1,11 @@
+import {StateTransitionEnum} from "../routes/enums/state.transition.type";
+
+const getTransitionTypeString = (id) => {
+    const [stateTransitionType] = Object.entries(StateTransitionEnum)
+        .filter(([key]) => StateTransitionEnum[key] === id)
+        .map(([key,]) => key)
+
+    return stateTransitionType ?? 'UNKNOWN'
+}
+
+export {getTransitionTypeString}
